+++ lib/src/state.rs
@@ -1,13 +1,9 @@
 use std::collections::BTreeMap;
-use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 
-use serde::{de, Deserializer};
 use serde_derive::{Deserialize, Serialize};
-use serde_json::Value;
 
 use crate::error::Error;
-use crate::settings::Settings;
 use crate::task::{Task, TaskStatus};
 
 pub const PUEUE_DEFAULT_GROUP: &str = "default";
@@ -54,73 +50,24 @@ pub struct Group {
 /// information, such as status changes and incoming commands by the client.
 #[derive(PartialEq, Eq, Clone, Debug, Deserialize, Serialize)]
 pub struct State {
-    /// The current settings used by the daemon.
-    pub settings: Settings,
     /// All tasks currently managed by the daemon.
     pub tasks: BTreeMap<usize, Task>,
     /// All groups with their current state a configuration.
-    #[serde(deserialize_with = "deserialize_groups")]
     pub groups: BTreeMap<String, Group>,
-    /// Used to store an configuration path that has been explicitely specified.
-    /// Without this, the default config path will be used instead.
-    pub config_path: Option<PathBuf>,
 }
 
-/// Custom group serializer, which tries to deserialize the field with the legacy representation if
-/// there are any errors. That way we can recover in a smooth way from the old format.
-/// This is necessary to ensure a semi-smooth transition from v1 to v2.
-/// TODO: Remove in 2.1.0
-fn deserialize_groups<'de, D>(deserializer: D) -> Result<BTreeMap<String, Group>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    // Do a general deserialization to Serde's `Value` type.
-    // That way we don't break the deserialization state if we're unable to deserialize it into
-    // our expected format.
-    let value: Value = serde::Deserialize::deserialize(deserializer)?;
-    let groups: Result<BTreeMap<String, Group>, serde_json::Error> =
-        serde_json::from_value(value.clone());
-
-    // If we cannot deserialize the state, this means that this is probably an old state which uses
-    // the old format. Try to deserialize the old format and convert them to the new format.
-    match groups {
-        Ok(groups) => Ok(groups),
-        Err(_) => {
-            let legacy_groups: Result<BTreeMap<String, GroupStatus>, serde_json::Error> =
-                serde_json::from_value(value);
-
-            let groups = match legacy_groups {
-                Ok(legacy_groups) => {
-                    // Iterate over all legacy groups and create a respective new group.
-                    let mut groups = BTreeMap::new();
-                    for (name, _) in legacy_groups.into_iter() {
-                        groups.insert(
-                            name,
-                            Group {
-                                status: GroupStatus::Paused,
-                                parallel_tasks: 1,
-                            },
-                        );
-                    }
-
-                    groups
-                }
-                Err(_) => return Err(de::Error::custom("Failed to deserialize `groups` field.")),
-            };
-
-            Ok(groups)
-        }
+impl Default for State {
+    fn default() -> Self {
+        Self::new()
     }
 }
 
 impl State {
     /// Create a new default state.
-    pub fn new(settings: &Settings, config_path: Option<PathBuf>) -> State {
+    pub fn new() -> State {
         let mut state = State {
-            settings: settings.clone(),
             tasks: BTreeMap::new(),
             groups: BTreeMap::new(),
-            config_path,
         };
         state.create_group(PUEUE_DEFAULT_GROUP);
         state
@@ -223,7 +170,7 @@ impl State {
     {
         // Return empty vectors, if there's no such group.
         if !self.groups.contains_key(group) {
-            return (vec![], vec![]);
+            return (Vec::new(), Vec::new());
         }
 
         // Filter all task ids of tasks that match the given group.